Update package.json; Small refactoring and renaming
