+++ frontend/src/common/ui-components/buttons/styles/ButtonWrapper.module.scss
@@ -1,5 +1,5 @@
-@import "../../../assets/styles/variables.scss";
-@import "../../../assets/styles/_base.scss";
+@import "../../../../assets/styles/variables.scss";
+@import "../../../../assets/styles/_base.scss";
 
 .button {
   font-size: 1em;
