+++ frontend/src/common/ui-components/common/Input/Input.module.scss
@@ -1,4 +1,4 @@
-@import "../../../assets/styles/_base";
+@import "../../../../assets/styles/_base";
 
 .inputWrapper {
   position: relative;
