+++ frontend/src/modules/Nodes/index.tsx
@@ -3,11 +3,11 @@ import { NodesContextProvider, useNodesContext } from "./context/NodesContext";
 // eslint-disable-next-line css-modules/no-unused-class
 import styles from "../Nodes/NodesPage.module.scss";
 import { NodeElementList } from "./components/NodeElement/NodeElementList";
-import PageName from "../../DEPRECATED_components/common/Page/PageName";
 import { RunningJob } from "./components/RunningJob/RunningJob";
 import { Results } from "./components/Results/Results";
 import BlueButton from "../../ui-lib/components/Button/BlueButton";
 import { SelectionContextProvider } from "../common/context/SelectionContext";
+import PageName from "../../common/ui-components/common/Page/PageName";
 
 const NodesPage = () => {
   const heading = "Run calibration node";
