+++ frontend/src/modules/SidebarMenu/styles/SidebarMenu.module.scss
@@ -28,7 +28,7 @@
   }
 }
 
-.entropyLogo {
+.qualibrateLogo {
   padding-left: $sidebar_padding;
   height: 50px;
   margin-bottom: 50px;
