+++ frontend/src/modules/WelcomePage/WelcomePage.module.scss
@@ -25,7 +25,7 @@
   overflow-y: scroll;
 }
 
-.entropyGuideFrame {
+.qualibrateGuideFrame {
   flex: 0.4;
 }
