+++ frontend/src/modules/common/Parameters/ParameterList.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { NodeDTO, NodeMap } from "../../Nodes/components/NodeElement/NodeElement";
 import { InputParameter, Parameters, SingleParameter } from "./Parameters";
 import { Checkbox } from "@mui/material";
-import InputField from "../../../DEPRECATED_components/common/Input/InputField";
 import { useGraphContext } from "../../GraphLibrary/context/GraphContext";
 import { GraphWorkflow } from "../../GraphLibrary/components/GraphList";
+import InputField from "../../../common/ui-components/common/Input/InputField";
 
 interface IProps {
   showParameters: boolean;
@@ -58,7 +58,7 @@ export const ParameterList: React.FC<IProps> = ({ showParameters = false, mapOfI
           <InputField
             placeholder={key}
             value={parameter.default ? parameter.default.toString() : ""}
-            onChange={(val) => {
+            onChange={(val: boolean | number | string) => {
               updateParameter(key, val, node);
             }}
           />
